# perf(collections/includes_value): support objects with non-enumerable properties