--- a/collections/includes_value.ts
+++ b/collections/includes_value.ts
@@ -23,7 +23,9 @@ export function includesValue<T>(
   record: Readonly<Record<string, T>>,
   value: T,
 ): boolean {
-  for (const i in record) {
+  const keys = getAllKeys(record);
+
+  for (const i of keys) {
     if (
       Object.hasOwn(record, i) &&
       (record[i] === value || Number.isNaN(value) && Number.isNaN(record[i]))
@@ -33,4 +35,22 @@ export function includesValue<T>(
   }
 
   return false;
+}
+
+/** Get all properties of an object (including non-enumerable properties) */
+function getAllKeys<T>(record: Readonly<Record<string, T>>) {
+  const keys: string[] = [];
+  let obj = record;
+
+  while (obj) {
+    Object.getOwnPropertyNames(obj).forEach((key) => {
+      if (keys.indexOf(key) === -1) {
+        keys.push(key);
+      }
+    });
+
+    obj = Object.getPrototypeOf(obj);
+  }
+
+  return keys;
 }