Since [`for...in` loop can only visit enumerable properties](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Enumerability_and_ownership_of_properties#traversing_object_properties), this PR updates `includesValue` function to work with non-enumerable properties.

And this update seems to make this function 15x - 35x faster.

Before:
```
benchmark      time (avg)             (min … max)       p75       p99      p995
------------------------------------------------- -----------------------------
test 1     34.45 ns/iter(28.12 ns … 467.91 ns) 29.44 ns 93.18 ns 109.36 ns
test 2     56.71 ns/iter(52.21 ns … 274.61 ns) 54.04 ns 111.51 ns 141.9 ns
test 3     78.49 ns/iter(74.45 ns … 245.08 ns) 75.48 ns 166.06 ns 177.85 ns
```
After:
```
benchmark      time (avg)             (min … max)       p75       p99      p995
------------------------------------------------- -----------------------------
test 1     2.4 µs/iter(1.6 µs … 5.59 ms) 1.9 µs 5.6 µs  8 µs
test 2     1.74 µs/iter(1.62 µs … 3.35 µs) 1.73 µs 3.35 µs 3.35 µs
test 3     2.3 µs/iter(1.7 µs … 4.44 µs) 2.34 µs 4.44 µs 4.44 µs
```