@babiabeo 2023-04-29
> Thanks for the PR! Are these results inverted? 1 microseconds is 1000 nanoseconds.

Oops! Sorry, my bad